requirements.txt
streamlit
google-generativeai
python-dotenv
notion-client
